Add freeboard to Monitoring
This is a great data viewer that's originally intended to easily display sensor data from IoT devices, but we have been using this in-house for a while to show disparate JSON data and it has been a joy to use. 

Check out the demos at http://freeboard.io/ and see for yourself!