remove GestióIP IPAM from Network Configuration Management
as I have recently migrated GestióIP from one host to another
and updated it from 3.2 to 3.4 in the process, I found that
it's no longer awesome because of the following reasons:

* it's hosted and supported on sourceforge
* the install script is >3500 lines of code
* you have to run said script with root
* there are no packages for any major distribution
* it's obfusc… er, written in perl
* the releases don't follow SemVer
* it does way more than IPAM, but other tools do it better
* it doesn't have a REST API (c'mon, it's $currentyear now)
* the weird XML/JSON API is only for the "pro version"
* there are lots of weird translations
* it uses cronjobs and doesn't know what systemd is
* it doesn't follow the Linux FHS
* there's no default config for nginx, only Apache
* did I mention the install script

I refuse to include software written like this on my list of
awesome sysadmin tools. You have to have fun using these tools,
they don't just have to "somehow work". The basic functionality
of GestióIP is there, I guess it's also used in countless legacy
installations worldwide, but I am of the opinion that many
of the design choices the (single?) GestióIP developer made are
not only weird, but also archaic and not awesome.

I recommend to use netbox, ralph or any of the other modern
solutions with a proper API and a proper FOSS development model.