remove Shinken (not awesome anymore)
sadly there has been no commit to the shinken repository
https://github.com/naparuba/shinken/commits/master
since 2019-07-16. There have been rumors that the project
is dead and the enterprise version hasn't seen many changes either.

For me this is a clear indication that the project isn't awesome
at this moment. Maybe this changes in the future but for now
I can't recommend Shinken for daily usage due to the fact that
there might be unfixed bugs or security holes and that it's hard
to find support anywhere else but the parent company.

If you're not okay with the removal of Shinken, please create a PR
and explain yourself.